requirements.txt
gunicorn==19.7.1
google-cloud-pubsub==0.30.1
google-cloud-bigquery==0.30.0